hotspots: Show hotspot only once for user profiles with same email.
It edits the function get_next_hotspot so only hotspots are shown
that have not been seen by the user or any other user profile with
the same email address.
Removes copy_hotspots because it is not needed any more.

Test cases for this new behaviour are added.

Manually tested by running a server and creating different users in
different realms.